+++ src/main/java/com/nisovin/shopkeepers/shopkeeper/offers/SKTradingOffer.java
@@ -11,6 +11,8 @@ import com.nisovin.shopkeepers.api.shopkeeper.TradingRecipe;
 import com.nisovin.shopkeepers.api.shopkeeper.offers.TradingOffer;
 import com.nisovin.shopkeepers.shopkeeper.SKTradingRecipe;
 import com.nisovin.shopkeepers.util.ItemUtils;
+import com.nisovin.shopkeepers.util.Log;
+import com.nisovin.shopkeepers.util.StringUtils;
 
 // shares its implementation with TradingRecipe, but always reports to not be out of stock
 public class SKTradingOffer extends SKTradingRecipe implements TradingOffer {
@@ -76,17 +78,25 @@ public class SKTradingOffer extends SKTradingRecipe implements TradingOffer {
 		}
 	}
 
-	public static List<SKTradingOffer> loadFromConfig(ConfigurationSection config, String node) {
+	public static List<SKTradingOffer> loadFromConfig(ConfigurationSection config, String node, String errorContext) {
 		List<SKTradingOffer> offers = new ArrayList<>();
 		ConfigurationSection offersSection = config.getConfigurationSection(node);
 		if (offersSection != null) {
 			for (String key : offersSection.getKeys(false)) {
 				ConfigurationSection offerSection = offersSection.getConfigurationSection(key);
-				if (offerSection == null) continue; // invalid offer: not a section
+				if (offerSection == null) {
+					// invalid offer: not a section
+					Log.warning(StringUtils.prefix(errorContext, ": ", "Invalid trading offer section for " + key));
+					continue;
+				}
 				ItemStack resultItem = offerSection.getItemStack("resultItem");
 				ItemStack item1 = offerSection.getItemStack("item1");
 				ItemStack item2 = offerSection.getItemStack("item2");
-				if (ItemUtils.isEmpty(resultItem) || ItemUtils.isEmpty(item1)) continue; // invalid offer
+				if (ItemUtils.isEmpty(resultItem) || ItemUtils.isEmpty(item1)) {
+					// invalid offer
+					Log.warning(StringUtils.prefix(errorContext, ": ", "Invalid trading offer for " + key + ": item1 or resultItem is empty"));
+					continue;
+				}
 				offers.add(new SKTradingOffer(resultItem, item1, item2));
 			}
 		}