+++ src/main/java/com/nisovin/shopkeepers/shopkeeper/player/sell/SKSellingPlayerShopkeeper.java
@@ -73,7 +73,7 @@ public class SKSellingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 	protected void loadFromSaveData(ConfigurationSection configSection) throws ShopkeeperCreateException {
 		super.loadFromSaveData(configSection);
 		// load offers:
-		this._setOffers(SKPriceOffer.loadFromConfig(configSection, "offers"));
+		this._setOffers(SKPriceOffer.loadFromConfig(configSection, "offers", "Shopkeeper " + this.getId()));
 	}
 
 	@Override