+++ src/main/java/com/nisovin/shopkeepers/shopkeeper/player/trade/SKTradingPlayerShopkeeper.java
@@ -65,7 +65,7 @@ public class SKTradingPlayerShopkeeper extends AbstractPlayerShopkeeper implemen
 	protected void loadFromSaveData(ConfigurationSection configSection) throws ShopkeeperCreateException {
 		super.loadFromSaveData(configSection);
 		// load offers:
-		this._setOffers(SKTradingOffer.loadFromConfig(configSection, "offers"));
+		this._setOffers(SKTradingOffer.loadFromConfig(configSection, "offers", "Shopkeeper " + this.getId()));
 	}
 
 	@Override